Introduce wind.py
This program can be run on RPi board which model is 3B+. To run this program you can get voltage data from sensor and calcute windspeed and finally send the value to remote server that is always listening. Also, the sent TCP information includes GPS information about latitude, longitude and UTC time.
The device also is connected to a LCD1602 display unit to display windspeed, position,time data in a loop.