introduce server.py
This program runs in a remote server with CentOS. In this server, there must be MySQL and python3.5 installed. The program can receive data from RPi through TCP connection and keep listening all the time. Then the reveived data will be stored in MySQL server to be queried or used to calaute and analyze later.